fix ordering for aggregate forecasts (#808)
cp change notification calculation used no ordering which caused an issue with finding the "current" CP value
this led to an edge case where an extreme prediction was compared to a within-bounds one which correctly calculates sotring difference but errored out on display difference. This is a needed fix, but ignores the math issue. Another PR will come that fixes the math.