comments: correct permissions for showing the edit option
until now, we were incorrectly showing 'edit' in the dropdown menu on
other people's comments when they were a reply to your comment.

this was because:
 - we passed down a single permissions object recursively which
   represented the amalgam of 'you can edit because you run the project'
and alseo 'you can edit because this was your comment'
 - we set that permission to an edit-allowing one any time a comment you
   create is found in the tree and never unset it (and indeed we
couldn't unset it, without also having the context re: post/project
permissions)

instead, we now let each comment determine its own editability directly.

Note that:

1. this adds the edit option to comments on your profile comment feed,
   which is a slight change of behaviour. I started to remove that again
but then thought maybe it's desirable to keep? so thought I'd ask first
2. I don't quite know enough of my way around to properly test the
   curator/admin side yet, but I've tested and the user side works (and
the code itself is pretty straightforward)

resolves #1160