comment dates: link to comment anchor
implements #666 (but not quite as requested)

AFAIK this approach -- having the timestamp be a link -- is the more
idiomatic approach, vs actually copying the link to the clipboard; I
think that it's preferable, as IMO it'll be more in line with user
expectations.

This is, for e.g., what Github does for its comments.