frontend/ci: add workflow checking prettier formatting (#1503)
ci: add workflow checking prettier formatting

see #1366

a few considerations:

1. I've set this up as a separate github workflow. I think lots of small
   workflows is the better pattern for these kinds of checks: it's quite
possible that with N different linters/checks, a bigger PR might fail on
multiple of them, which leads to annoying fix-push-check pingpong if you
have several checks in sequence in a regular action. Yes you can write
the action itself to parallelize but that can end up surprisingly
complex - plus, with caching, the overhead of parallel workflows can be
quite small.

2. To do this practically, you need to pin the exact version of
   prettier/its plugins, as by definition different versions format ever
so slightly differently. Note that I've pinned not quite the latest
prettier (3.3.1 not 3.3.3) because empirically that appears to be what
people are using (3.3.3 wants to change quite a few files).

This may take a tiny bit of editor config at first, if some folks are
using a global prettier version not the project prettier by accident -
but most editors detect this by default.

3. As a corollary of (2), you'll now need to run `npm run format` for
   every prettier/plugin version bump or the build will fail.

Co-authored-by: Hlib <hlib@metaculus.com>