Show a box with two actions after the user predicts
One of the buttons is the Follow button which is reused. And the other
button is getting the user to the comment editor box and checks the box
for including the prediction in the comment. I chose to do this via an
URL param (which gets cleared right after) to avoid the overkill of
passing props or using a context.